requirements.txt
# MAXCAPITAL Telegram Bot - Dependencies
# Python 3.10+

# Telegram Bot Framework
aiogram==3.3.0
aiohttp==3.9.1

# Database
asyncpg==0.29.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4

# OpenAI
openai==1.6.1

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==23.2.0
colorama==0.4.6

# Utilities
aiofiles==23.2.1
httpx==0.25.2

